+++ samples/Clients/old/Directory.Build.props
@@ -3,9 +3,7 @@
 	 <Import Project="../Directory.Build.props" />
   </ImportGroup>
   <ItemGroup>
+      <Compile Include="$(SolutionDir)MvcUsings.cs" Link="MvcUsings.cs" />
      <ProjectReference Include="$(SolutionDir)..\shared\Constants\Constants.csproj" Condition="$(MSBuildProjectName) != 'Constants' " />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="$(SolutionDir)MvcUsings.cs" Link="MvcUsings.cs" />
   </ItemGroup>
 </Project>
