+++ samples/Clients/src/Directory.Build.props
@@ -3,6 +3,6 @@
 	 <Import Project="../Directory.Build.props" />
   </ImportGroup>
   <ItemGroup>
-     <ProjectReference Include="$(SolutionDir)\shared\Constants\Constants.csproj" Condition="$(MSBuildProjectName) != 'Constants' " />
+     <ProjectReference Include="$(SolutionDir)..\shared\Constants\Constants.csproj" Condition="$(MSBuildProjectName) != 'Constants' " />
   </ItemGroup>
 </Project>
