+++ samples/KeyManagement/Directory.Build.props
@@ -1,15 +1,14 @@
 <Project>
-  <ImportGroup>
-	 <Import Project="../Directory.Build.props" />
-  </ImportGroup>
+    <ImportGroup>
+        <Import Project="../Directory.Build.props" />
+    </ImportGroup>
   <ItemGroup>
-      <PackageReference Include="IdentityServer4.KeyManagement" />
-      <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
-      <PackageReference Include="Microsoft.EntityFrameworkCore.Design" >
-          <PrivateAssets>all</PrivateAssets>
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      </PackageReference>
-      <Compile Remove="$(SolutionDir)..\..\SolutionUsings.cs" />
-	<Compile Include="$(SolutionDir)\..\SolutionUsings.cs" Link="SolutionUsings.cs" />
-  </ItemGroup>
+    <PackageReference Include="IdentityServer4.KeyManagement" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" >
+	   <PrivateAssets>all</PrivateAssets>
+	   <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <Compile Include="$(SolutionDir)..\KeyManagementGlobalUsings.cs" Link="KeyManagementGlobalUsings" />
+  </ItemGroup>	
 </Project>
