+++ samples/KeyManagement/FileSystemKeys/Directory.Build.props
@@ -1,0 +1,5 @@
+<Project>
+    <ImportGroup>
+        <Import Project="../Directory.Build.props" />
+    </ImportGroup>
+</Project>
