+++ samples/KeyManagement/FileSystemKeys/FileSystem/Properties/launchSettings.json
