+++ samples/KeyManagement/FileSystemKeys/FileSystem/Startup.cs
@@ -9,7 +9,6 @@
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
 */
-
 public class Startup
 {
     public IWebHostEnvironment Environment { get; }
