+++ samples/KeyManagement/FileSystemKeys/database/EF/Config.cs
@@ -10,29 +10,24 @@
  copies or substantial portions of the Software.
 */
 
-using IdentityServer4.Models;
-using System.Collections.Generic;
-
-namespace sample
+public static class Config
 {
-    public static class Config
+    public static IEnumerable<IdentityResource> GetIdentityResources()
     {
-        public static IEnumerable<IdentityResource> GetIdentityResources()
+        return new IdentityResource[]
         {
-            return new IdentityResource[]
-            {
                 new IdentityResources.OpenId()
-            };
-        }
+        };
+    }
 
-        public static IEnumerable<ApiResource> GetApis()
-        {
-            return new ApiResource[] { };
-        }
+    public static IEnumerable<ApiResource> GetApis()
+    {
+        return new ApiResource[] { };
+    }
 
-        public static IEnumerable<Client> GetClients()
-        {
-            return new Client[] { };
-        }
+    public static IEnumerable<Client> GetClients()
+    {
+        return new Client[] { };
     }
 }
+
