+++ samples/KeyManagement/FileSystemKeys/database/EF/EfSample.csproj
