+++ samples/KeyManagement/FileSystemKeys/database/EF/Properties/launchSettings.json
