--- samples/KeyManagement/global.json.old
@@ -1,6 +1,0 @@
-{
-  "sdk": {
-    "version": "3.1.100",
-    "rollForward": "latestPatch"
-  }
-}
